das-92 #comment WIP - refactoring the ContractData handler
