+++ hat/conf/v26.routes
@@ -79,6 +79,6 @@ GET           /she/static/$endpoint<[0-9a-z-/]+>                                
 
 
 # Contract Data
-POST          /contract-data/read/$namespace<[0-9a-z-]+>/$endpoint<[0-9a-z-/]+>              org.hatdex.hat.api.controllers.RichData.getContractData(namespace, endpoint, orderBy: Option[String], ordering: Option[String], skip: Option[Int], take: Option[Int])
-POST          /contract-data/create/$namespace<[0-9a-z-]+>/$endpoint<[0-9a-z-/]+>            org.hatdex.hat.api.controllers.RichData.saveContractData(namespace, endpoint, skipErrors: Option[Boolean])
-PUT           /contract-data/update/$namespace<[0-9a-z-]+>/$endpoint<[0-9a-z-/]+>            org.hatdex.hat.api.controllers.RichData.updateContractData(namespace, endpoint, skipErrors: Option[Boolean])
+POST          /contract-data/read/$namespace<[0-9a-z-]+>/$endpoint<[0-9a-z-/]+>              org.hatdex.hat.api.controllers.ContractData.getContractData(namespace, endpoint, orderBy: Option[String], ordering: Option[String], skip: Option[Int], take: Option[Int])
+POST          /contract-data/create/$namespace<[0-9a-z-]+>/$endpoint<[0-9a-z-/]+>            org.hatdex.hat.api.controllers.ContractData.saveContractData(namespace, endpoint, skipErrors: Option[Boolean])
+PUT           /contract-data/update/$namespace<[0-9a-z-]+>/$endpoint<[0-9a-z-/]+>            org.hatdex.hat.api.controllers.ContractData.updateContractData(namespace, endpoint, skipErrors: Option[Boolean])
